removed axes
