+++ Python/tdw/add_ons/oculus_touch.py
@@ -1,4 +1,5 @@
 from typing import List, Callable, Dict
+import numpy as np
 from tdw.add_ons.autohand import Autohand
 from tdw.vr_data.oculus_touch_button import OculusTouchButton
 from tdw.vr_data.rig_type import RigType
@@ -38,6 +39,9 @@ class OculusTouch(Autohand):
                          avatar_camera_width=avatar_camera_width, headset_aspect_ratio=headset_aspect_ratio,
                          headset_resolution_scale=headset_resolution_scale, non_graspable=non_graspable,
                          discrete_collision_detection_mode=discrete_collision_detection_mode)
+        # Axis events.
+        self._axis_events_left: List[Callable[[np.array], None]] = list()
+        self._axis_events_right: List[Callable[[np.array], None]] = list()
         # Button press events.
         self._button_press_events_left: Dict[OculusTouchButton, Callable[[], None]] = dict()
         self._button_press_events_right: Dict[OculusTouchButton, Callable[[], None]] = dict()
@@ -81,6 +85,19 @@ class OculusTouch(Autohand):
             self._button_press_events_left[button] = function
         else:
             self._button_press_events_right[button] = function
+
+    def listen_to_axis(self, is_left: bool, function: Callable[[np.array], None]) -> None:
+        """
+        Listen for Oculus Touch controller axis events.
+
+        :param is_left: If True, this is the left controller. If False, this is the right controller.
+        :param function: The function to invoke when the button is pressed. This function must a single argument (a numpy array of shape `(2)`, representing (x, y) coordinates) and return None.
+        """
+
+        if is_left:
+            self._axis_events_left.append(function)
+        else:
+            self._axis_events_right.append(function)
 
     def _get_human_hands(self) -> RigType:
         return RigType.oculus_touch_human_hands
